pyproject.toml
[project]
name = "sender"
version = "0.1.0"
description = "Telegram bot for bulk messaging"
authors = [
    {name = "Honley"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
